Migrate data providers to state providers
